Issue 2 item index

+++ db/schema.rb
@@ -22,7 +22,7 @@ ActiveRecord::Schema.define(version: 20170222010254) do
   create_table "items", force: :cascade do |t|
     t.string   "title"
     t.string   "description"
-    t.decimal  "price",       precision: 2
+    t.float    "price"
     t.string   "image_tag"
     t.datetime "created_at",                null: false
     t.datetime "updated_at",                null: false
